+++ src/middlewares/storage.ts
@@ -2,6 +2,7 @@ import * as fs from 'fs';
 import * as path from 'path';
 import mkdirSync from '../utils/mkdir-sync';
 import { Middleware } from '..';
+import { ISourceContext } from '../utils/source';
 import { IFetchContext } from './fetch';
 
 
@@ -35,7 +36,7 @@ declare module storage {
 
 const storage = {
 
-    file(options: storage.IFileOptions = {}): Middleware<{}, IFetchContext & storage.IFileContext> {
+    file(options: storage.IFileOptions = {}): Middleware<{}, ISourceContext & IFetchContext & storage.IFileContext> {
 
         return async function (ctx, next) {
             // after `fetch` before `router`
@@ -49,8 +50,8 @@ const storage = {
             storageFileBySteam(ctx) || storageFileByString(ctx);
         }
 
-        function storageFileBySteam({ req: { file }, res }: IFetchContext & storage.IFileContext) {
-            if (file == null || res == null || res.bodyUsed) return false;
+        function storageFileBySteam({ state: { file, content }, res }: IFetchContext & storage.IFileContext) {
+            if (file == null || content != null || res == null || res.bodyUsed) return false;
 
             mkdirSync(path.dirname(file), options);
             res.body.pipe(fs.createWriteStream(path.join(options.root || process.cwd(), file)));